requirements.txt
fastapi==0.103.2
uvicorn[standard]==0.23.2
python-multipart==0.0.6
Pillow==10.0.1
torch==2.1.0
transformers==4.34.0
pydantic==2.4.2
